Add TODO for catching errors thrown by sub-dependency resolution
Problem: if two neighboring dependencies each depend on incompatible versions of a package, this incompatibility isn't discovered until the reduction phase, by which point it's too late to go back and re-resolve. Even if we wanted to correct this incompatibility, we end up with the problem below:

A -> C >2.0
B -> C <2.0

To resolve this, do we try a lower version of A or a lower version of B? We have to choose one. Possible strategy: it's reasonable to assume that people tend to prefer newer versions of packages, and it's probably less likely that a package will update to a new version with an older dependency, so we can downgrade packages that have a higher dependency version requirement, since they're more likely to have older versions of their dependencies as their version decreases. We want to downgrade as little as possible, so we pick the packages that have dependencies with the highest possible version to downgrade first. In the situation above, 'A' would be chosen to downgrade since it is more likely to have previously depended on an older version of 'C'. This might seem obvious, but it's only based on a reasonable assumption, so it is a tradeoff that could cause resolution issues for some packages.

The next problem is how we detect these incompatibilities in a multithreaded context, since both of the neighboring dependencies are being resolved in parallel.